+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for calling constructors functions with NewInstance on Function
+
 ## [v0.6.0] - 2021-05-11
 
 ### Added